refactor: overhaul project structure
