+++ internal/discord_frontend/abort.go
@@ -1,10 +1,10 @@
-package appcmd
+package discord_frontend
 
 import (
 	"github.com/bwmarrin/discordgo"
-	"github.com/mmm-development/bulker-discord/bend"
-	"github.com/mmm-development/bulker-discord/clog"
+	"github.com/mmm-development/bulker-discord/internal/clog"
 	"github.com/mmm-development/bulker-discord/locale"
+	bend "github.com/mmm-development/bulker-discord/pkg/bulker_backend"
 )
 
 var (
